requirements.txt
Django==1.8.14
djangorestframework==3.6.4
django-cors-headers==2.2.0
redis==2.10.6
django-websocket-redis==0.5.2
celery==3.1.23
